+++ src/server/features/completion/completion-context.ts
