+++ src/server/features/completion/completion-utils.ts
