+++ src/server/features/completion/function-completion.ts
