+++ src/server/features/completion/variable-completion.ts
