+++ src/server/features/diagnostics.ts
