+++ src/server/features/workspace-symbol.ts
