+++ src/test/features/completion.spec.ts
@@ -3,8 +3,8 @@ import { CompletionItemKind, SymbolKind } from "vscode-languageserver";
 import type { CompletionList } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
 import { ScssDocument } from "../../server/document";
-import { doCompletion } from "../../server/providers/completion";
-import { parseStringLiteralChoices } from "../../server/providers/completion/completion-utils";
+import { doCompletion } from "../../server/features/completion";
+import { parseStringLiteralChoices } from "../../server/features/completion/completion-utils";
 import { sassBuiltInModules } from "../../server/sass-built-in-modules";
 import StorageService from "../../server/services/storage";
 import type { ISettings } from "../../server/types/settings";
