+++ src/test/features/diagnostics.spec.ts
@@ -2,7 +2,7 @@ import assert from "assert";
 import fs, { Stats } from "fs";
 import { stub, SinonStub } from "sinon";
 import { DiagnosticSeverity, DiagnosticTag } from "vscode-languageserver-types";
-import { doDiagnostics } from "../../server/providers/diagnostics";
+import { doDiagnostics } from "../../server/features/diagnostics";
 import StorageService from "../../server/services/storage";
 import * as fsUtils from "../../server/utils/fs";
 import { EXTENSION_NAME } from "../../shared/constants";
