+++ src/test/features/go-definition.spec.ts
@@ -2,7 +2,7 @@ import assert from "assert";
 import { SymbolKind } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
 import { ScssDocument } from "../../server/document";
-import { goDefinition } from "../../server/providers/go-definition";
+import { goDefinition } from "../../server/features/go-definition";
 import StorageService from "../../server/services/storage";
 import * as helpers from "../helpers";
