+++ src/test/features/references.spec.ts
@@ -1,7 +1,7 @@
 import assert from "assert";
 import fs, { Stats } from "fs";
 import { stub, SinonStub } from "sinon";
-import { provideReferences } from "../../server/providers/references";
+import { provideReferences } from "../../server/features/references";
 import StorageService from "../../server/services/storage";
 import * as fsUtils from "../../server/utils/fs";
 import * as helpers from "../helpers";
