+++ src/test/features/signature-help.spec.ts
@@ -3,7 +3,7 @@ import { SymbolKind } from "vscode-languageserver";
 import type { SignatureHelp } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
 import { ScssDocument } from "../../server/document";
-import { doSignatureHelp } from "../../server/providers/signature-help";
+import { doSignatureHelp } from "../../server/features/signature-help";
 import StorageService from "../../server/services/storage";
 import * as helpers from "../helpers";
